challenge solved
It is a chllenge from React.dev learning curriculum.
Problem was to extract a component out of the provided codebase to reduced duplicity
generate an array to store all the props data and created a Profile component.
Can use map method to loop over the array to generate Profile component.