optimized accordion with 'lifting the state up' concept
implement the state control functionality to Accordion component
pass isActive prop to Panel component which will return either true or false.
pass an onShow prop to Panel component which will be invoked on click of a button.
now changing the state of the child components(Panel) is the responsibility of parent component(Accordion).